feat(apps/prod): add tiup-publisher (#1280)
This pull request adds a new service, `tiup-publisher`, to the
production environment. The changes include updates to the kustomization
files and the addition of HelmRelease configurations for both production
and staging mirrors.

### Additions to `tiup-publisher` service:

*
[`apps/prod/kustomization.yaml`](diffhunk://#diff-ee20854a69c2a0d1fcc07e8216c968e5895b76a9f9ffc942a8aa69a9747354e5R25):
Added `tiup-publisher` to the list of resources.
*
[`apps/prod/tiup-publisher/kustomization.yaml`](diffhunk://#diff-54630566ef0b0b3c9dcdcd9d06b3bb924fc49d8a20364ea78c06c21a8fea5cfbR1-R5):
Created a new kustomization file for `tiup-publisher` with references to
`release-prod-mirror.yaml` and `release-staging-mirror.yaml`.

### HelmRelease configurations:

*
[`apps/prod/tiup-publisher/release-prod-mirror.yaml`](diffhunk://#diff-6e5ab966061a7a1c3ec666fc99421d86aa4ce2aba340522ff22e909439fc733cR1-R47):
Added HelmRelease configuration for the production mirror of
`tiup-publisher`, including image tag, volumes, and node selector.
*
[`apps/prod/tiup-publisher/release-staging-mirror.yaml`](diffhunk://#diff-4cd3cadac6a6bd62d698da77ccc6503a663efced3f2319f01ac0fdbb7fa5fec4R1-R47):
Added HelmRelease configuration for the staging mirror of
`tiup-publisher`, including image tag, volumes, and node selector.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>

Signed-off-by: wuhuizuo <wuhuizuo@126.com>