fix(apps/prod/tekton/configs/tasks/release): avoid label be removed by prow plugins (#1286)
This pull request includes a small but important change to the
`create-pr-to-bump-tikv-version.yaml` file in the `spec` directory. The
change addresses an issue where the prow plugin removes the approved
label by adding a delay and reapplying the necessary labels.

*
[`apps/prod/tekton/configs/tasks/release/create-pr-to-bump-tikv-version.yaml`](diffhunk://#diff-a96c12396451377af34efa48805fc711eed0d71f0d6b51a4c6f94a0fa91edb85R128-R133):
Added a sleep command and logic to reapply the `lgtm`, `approved`, and
`cherry-pick-approved` labels to the pull request after creation.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>

Signed-off-by: wuhuizuo <wuhuizuo@126.com>