feat(apps/prod): deploy chatops lark bot (#1427)
This pull request introduces a new HelmRelease for the `chatops-lark`
application and updates the kustomization files to include this new
resource. The most important changes include the addition of
`chatops-lark` to the kustomization resources and the detailed
configuration of the HelmRelease.

Additions to kustomization and HelmRelease configuration:

*
[`apps/prod/chatops-lark/kustomization.yaml`](diffhunk://#diff-4119b17540b0e13598661402ee1efba2b16ef8512cab59ab1f20a80bc40646eaR1-R5):
Created a new kustomization file for the `chatops-lark` application.
*
[`apps/prod/chatops-lark/release.yaml`](diffhunk://#diff-20c9387c8f1e6985fb4d41f78c9b9b82835b612abbd3e778c45947b2de277079R1-R46):
Added detailed HelmRelease configuration for `chatops-lark`, including
chart specifications, installation and upgrade strategies, resource
limits, and image details.
*
[`apps/prod/kustomization.yaml`](diffhunk://#diff-ee20854a69c2a0d1fcc07e8216c968e5895b76a9f9ffc942a8aa69a9747354e5R27):
Included `chatops-lark` in the list of resources to be managed by
kustomize.

Signed-off-by: wuhuizuo <wuhuizuo@126.com>