fix(apps/gcp/prow): finish migrate for gateways into infra (#1534)
This pull request removes configurations related to the `gateways`
component in the `apps/gcp` directory and updates references to the
`external-https` Gateway to use the `infra` namespace instead of `apps`.
The most important changes are grouped below by theme.

### Removal of `gateways` component:

*
[`apps/gcp/gateways.yaml`](diffhunk://#diff-4c53ea98d89be42fe6c2d76ce90425fcf6a459fda5b298a7cb1617e2f12f0ca7L1-L17):
Deleted the `Kustomization` configuration for the `gateways` component,
including its source reference and post-build substitution for the
`INGRESS_STATIC_IP_NAME`.
*
[`apps/gcp/gateways/gke-gateway.yaml`](diffhunk://#diff-d2142dbf4653658f890b1c45a4663ede283a9e3bcf90d6b5834deebddc4d273eL1-L23):
Removed the `Gateway` resource configuration for `external-https`, which
defined HTTPS listeners and static IP addresses.
*
[`apps/gcp/gateways/kustomization.yaml`](diffhunk://#diff-e08531d6d0a10531f604dcb010bcb110e6c619141b5239afa8b8bd6ef192fbb5L1-L5):
Deleted the `Kustomization` file that referenced `gke-gateway.yaml`.
*
[`apps/gcp/kustomization.yaml`](diffhunk://#diff-1a6255cbbcce18a6d6b8cd1979969f2f22613b4d051c887c3c23e4cb06ceef63L6):
Removed the reference to `gateways.yaml` from the `resources` list in
the `Kustomization` file for the `apps/gcp` directory.

### Namespace update for Gateway references:

*
[`apps/gcp/prow/release/http-routes.yaml`](diffhunk://#diff-cf0eb49a5180fad738bce9b769f1ad3c26161994f95ac59371f3d851b62ea127L10-R10):
Updated the `namespace` field for the `external-https` Gateway in the
`parentRefs` section from `apps` to `infra`.