feat(apps/prod2): add zot deployment manifests for prod2 environment (#1569)
This pull request introduces configuration files for deploying the `zot`
application in the `prod2` environment using Kubernetes and FluxCD. It
includes the creation of a namespace, HelmRepository, and HelmRelease
resources for `zot` and its dependency, `redis`. The changes ensure
proper resource management, ingress configuration, and persistence
settings for the application.

### Kubernetes Configuration:

*
[`apps/prod2/zot/kustomization.yaml`](diffhunk://#diff-157989a6caf6e2e2f6ac034ed289e893a435f5b3fb5b60352dc1b01108c2c26dR1-R8):
Added a `Kustomization` file to manage resources for the `zot`
application, including namespace, source, and HelmRelease
configurations.
*
[`apps/prod2/zot/namespace.yaml`](diffhunk://#diff-29fee16c514bf40b4d83cef58d10ab9e9f91c666aa010b03b65b83dcff29cf34R1-R6):
Created a `Namespace` resource named `zot` with annotations specifying
node selector constraints for scheduling.

### HelmRepository Setup:

*
[`apps/prod2/zot/source.yaml`](diffhunk://#diff-798ccfcd5255931424b37bf2446131fac91f19c6adc1a0604e26464e2656bc7bR1-R7):
Defined a `HelmRepository` resource for `zot` with a URL pointing to the
`zotregistry.dev` Helm charts repository and a sync interval of 24
hours.

### Application Deployment:

*
[`apps/prod2/zot/release.yaml`](diffhunk://#diff-2e556271ae8b53b6860d3146c3fe9945eff9dec998cf654312fefee233672de8R1-R43):
Configured a `HelmRelease` for the `zot` application, specifying chart
details, dependencies (on `redis`), ingress settings, persistence
options, and service type.

### Redis Dependency:

*
[`apps/prod2/zot/redis-release.yaml`](diffhunk://#diff-db1e54e61b32d14b7989a1b231065be7ab5a2f162124704f24034be245115d83R1-R22):
Added a `HelmRelease` for the `redis` dependency, with resource presets
set to `nano` and storage class configured for single replica.