feat(gcp): do not run application pods on spot nodes (#1591)
This pull request introduces a Kyverno policy to prevent application
pods from running on GKE spot nodes, integrates Kyverno into the
infrastructure, and updates various kustomization files to reflect these
changes. The most important changes include adding the Kyverno Helm
repository, creating a HelmRelease for Kyverno, and defining the new
cluster policy.

### Kyverno Integration:

*
[`infrastructure/_base/sources/helm-repo-kyverno.yaml`](diffhunk://#diff-e0904f0ce6dac140ec3c2a3f413d4b6d2ebd5058ec02768d7ca5e97cfee5ca2aR1-R8):
Added a Helm repository definition for Kyverno with a 10-minute sync
interval.
*
[`infrastructure/gcp/kyverno/release.yaml`](diffhunk://#diff-249fed7a1d6301dca520894988224b6148c379123b57cd7a8bc53c47519fb2cbR1-R25):
Added a HelmRelease for deploying Kyverno in the `infra` namespace with
specific replica configurations for various controllers.
*
[`infrastructure/gcp/kyverno/kustomization.yaml`](diffhunk://#diff-fd04d4dc54d43e59835b9cb3b666fc2119c767c94378d8313a5999b020aa1c5fR1-R4):
Added a kustomization file to manage the Kyverno release.
*
[`infrastructure/_base/sources/kustomization.yaml`](diffhunk://#diff-c2e7d0b3ba9f2e6b9078c6f780dd493ace75795751559417fca59aa9c4de73fdR10):
Included the Kyverno Helm repository in the list of resources.
*
[`infrastructure/gcp/kustomization.yaml`](diffhunk://#diff-8054653b8a00c5bf4116eaca553d7c7f7334d5543e9557211fedc65465cf8590R8):
Added Kyverno to the GCP-specific kustomization resources.

### New Cluster Policy:

*
[`apps/gcp/cluster-policies/avoid-application-pods-run-on-spot-nodes.yaml`](diffhunk://#diff-50897996062e8c105b6855f52378a3f15fe73066f5b5af85c25ba95921b658c7R1-R29):
Created a Kyverno ClusterPolicy to prevent application pods from running
on GKE spot nodes using a node affinity rule.
*
[`apps/gcp/cluster-policies/kustomization.yaml`](diffhunk://#diff-31a41f8634499693bb9ce2cf7dbbe23f54bc7198d5392b7b46abef21533154fbR1-R4):
Added a kustomization file to manage the new cluster policy.

### Kustomization Updates:

*
[`apps/gcp/kustomization.yaml`](diffhunk://#diff-1a6255cbbcce18a6d6b8cd1979969f2f22613b4d051c887c3c23e4cb06ceef63L5-R5):
Replaced `namespace.yaml` with `cluster-policies` in the list of
resources.
*
[`apps/gcp/prow/kustomization.yaml`](diffhunk://#diff-2b6196b4b90ee56313b9042005841aa78ecb05287bb4143a918da7818c1c1475R4):
Added `namespace.yaml` to the Prow kustomization resources.

### Dependency Management:

*
[`apps/gcp/prow/release.yaml`](diffhunk://#diff-dde7f5b4877c12b9029362129d12ad53cac7f72a297820f76694f2a06bf4aa92R10-R11):
Added a dependency on `prow-crd` for the Prow release.
*
[`clusters/gcp/apps.yaml`](diffhunk://#diff-0a92019f5f0ac95940040457f1e64f7205d0839459e67de76c6dc33d9ff042c7L1-R9):
Added a dependency on `infrastructure` for the apps kustomization.