feat(apps/prod2): add cron jobs to fetch for branch image tags (#1747)
This pull request introduces updates to the Tekton dashboard extension
configuration and adds a new post-deployment Kustomization with a
scheduled CronJob for the `zot` application in the `prod2` environment.
The changes improve resource organization and add automation for
managing image tags.

**Tekton Dashboard Extension Updates:**
- Added the `namespace: tekton-pipelines` field to the `cronjobs`
extension metadata for both `gcp` and `prod2` environments to ensure
proper namespacing.

**Zot Application Post-Deployment Automation:**
- Added a new `post.yaml` Kustomization resource to the `zot` app, which
depends on `zot-pre` and deploys resources from the `post` directory.
[[1]](diffhunk://#diff-f97c6045637160497f033dfea3c18763812d78f6540ce02adcd878c185bed2e9R1-R15)
[[2]](diffhunk://#diff-157989a6caf6e2e2f6ac034ed289e893a435f5b3fb5b60352dc1b01108c2c26dR10)
- Created a new CronJob resource,
`cronjob-fetch-branch-tags-mirror-tidbx`, scheduled to run hourly. This
job attempts to update and verify specific image tags for several
repositories, aiding in image mirroring and tag management.
- Added a `kustomization.yaml` in the `post` directory to manage the new
CronJob resource.